+++ ignite/cmd/chain_faucet.go
@@ -43,7 +43,7 @@ func chainFaucetHandler(cmd *cobra.Command, args []string) error {
 		chain.CollectEvents(session.EventBus()),
 	}
 
-	c, err := newChainWithHomeFlags(cmd, chainOption...)
+	c, err := NewChainWithHomeFlags(cmd, chainOption...)
 	if err != nil {
 		return err
 	}